{{cookiecutter.plugin_name}}/pyproject.toml
[tool.poetry]
name = "{{cookiecutter.plugin_name}}"
version = "0.1.0"
description = ""
authors = ["<your team name> <<your team's email address>>"]
readme = "README.md"
packages = [{include = "plugin"}]

[tool.poetry.dependencies]
python = "^3.9,<3.13"
cloudquery-plugin-sdk = "^0.1.24"
pydantic = "^2.4"
pyarrow = "^15.0.2"
requests = "^2.32.0"
botocore = "^1.34.26"
python_dateutil = "^2.5.3"
typing-extensions = "^4.7.1"
structlog = "^23.1.0"

[tool.poetry.scripts]
main = "main:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
